+++ caladrius/interface/client/src/map-widget/Map.js
@@ -6,14 +6,9 @@ import {
     LayerGroup,
     LayersControl,
 } from "react-leaflet";
-import HeatmapLayer from "react-leaflet-heatmap-layer";
 import "leaflet/dist/leaflet.css";
 import "./map.css";
-import {
-    get_prediction_colour,
-    contrast_color_array,
-    get_heatmap_gradient,
-} from "../colours";
+import { get_prediction_colour, contrast_color_array } from "../colours";
 
 const MAPBOX_ACCESS_TOKEN =
     "pk.eyJ1IjoiZ3VsZmFyYXoiLCJhIjoiY2p6NW10bmxhMGRidzNldDQ1ZmwxZ2gwbCJ9.tqPa766Wzm0xwy0p9_T3Jg";
@@ -26,16 +21,10 @@ const INTERACTION_ZOOM_LEVEL = 18;
 let DEFAULT_CENTER_COORDINATES = [18.035, -63.07];
 
 export class Map extends React.Component {
-    constructor(props) {
-        super(props);
-        this.get_building_shape_array = this.get_building_shape_array.bind(
-            this
-        );
-        this.get_admin_regions = this.get_admin_regions.bind(this);
-    }
-
-    get_building_shape_array() {
-        let building_shape_array = this.props.data.map(datum => {
+    get_building_shape_array(key, label) {
+        let building_shape_array = (
+            this.props.data[key].slice(-1)[0] || []
+        ).map(datum => {
             const colour = get_prediction_colour(
                 datum.prediction,
                 this.props.damage_boundary_a,
@@ -64,7 +53,7 @@ export class Map extends React.Component {
             );
         });
         return (
-            <LayersControl.Overlay name={"Buildings"} checked={true}>
+            <LayersControl.Overlay name={label} checked={true}>
                 <LayerGroup>{building_shape_array}</LayerGroup>
             </LayersControl.Overlay>
         );
@@ -112,31 +101,6 @@ export class Map extends React.Component {
         );
     }
 
-    prediction_heat_map(cacheData) {
-        let heatCoordinates = cacheData.map(x => [
-            x.coordinates[0][0],
-            x.coordinates[0][1],
-            x.prediction,
-        ]);
-        return (
-            <LayersControl.Overlay name={"Heat Map"} checked={true}>
-                <HeatmapLayer
-                    points={heatCoordinates}
-                    longitudeExtractor={m => m[1]}
-                    latitudeExtractor={m => m[0]}
-                    intensityExtractor={m => parseFloat(m[2])}
-                    minOpacity={0.2}
-                    radius={20}
-                    blur={5}
-                    gradient={get_heatmap_gradient(
-                        this.props.damage_boundary_a,
-                        this.props.damage_boundary_b
-                    )}
-                />
-            </LayersControl.Overlay>
-        );
-    }
-
     render() {
         const center_coordinates = this.props.selected_datum
             ? this.props.selected_datum.coordinates[0]
@@ -162,8 +126,15 @@ export class Map extends React.Component {
                         MAPBOX_ACCESS_TOKEN
                     )}
                     {this.get_admin_regions()}
-                    {this.prediction_heat_map(this.props.data)}
-                    {this.get_building_shape_array()}
+                    {this.get_building_shape_array(
+                        "validation",
+                        "Validation Set"
+                    )}
+                    {this.get_building_shape_array("test", "Test Set")}
+                    {this.get_building_shape_array(
+                        "inference",
+                        "Inference Set"
+                    )}
                 </LayersControl>
             </LeafletMap>
         );