+++ caladrius/interface/client/src/scatter-plot/Circles.js
@@ -38,8 +38,13 @@ class Circle extends React.Component {
 
     draw_plot() {
         const maximum_opacity = 1.0;
-        const minimum_opacity =
-            this.props.selected_datum === this.props.datum ? 1.0 : 0.3;
+        let minimum_opacity = 0.3;
+        if (
+            this.props.selected_datum &&
+            this.props.selected_datum.object_id === this.props.datum.object_id
+        ) {
+            minimum_opacity = maximum_opacity;
+        }
         d3.select(this.ref.current)
             .attr("cx", this.props.axis.x_scale(this.props.datum.prediction))
             .attr("cy", this.props.axis.y_scale(this.props.datum.label))